requirements.txt
Office365-REST-Python-Client==2.5.13
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1